Animation, Models & stimuli
Added a heart beat animation and adjustability to the rate. Edited scale of models to match, though still needs to be fixed as new scene seemed to mess with prefab scaling. Adding in all the new assets for the first build. Working towards fixing some errors with the stimulus system and applying animations