Stimulus, Selection Ring & Arterioles
Stimulus system refined to be more usable in editor when adding new cells, organs and the various ways they can respond.

Added in simple Arterioles and created a simple template tissue structure at the first view.

Created a growing and shrinking animation for the selection ring and used the Animator Controlled to control when it does either.