Cell Reset, Saving Materials & Distance
Successfully implemented distance calculation, stimulus only affects cells within given distance from centre.

Edited Co-ordinate system, organ & stimulus to save a reference of the new materials of the cell if it's changed and to reset the cell objects in the pool with their default materials.

Adding in system for deleting co-ordinates of dead cells but issues some issues with instanciating a new array/list.